Added max patch, to begin analyzing the data output by CAMusic.java.
I'm adding in some basic features for acquiring useful information from CAMusic. CAMusic will be able to report about cellular density for a given state. To do this, it will maintain an array of how many cells are in each state. For some reason, I'm getting negative numbers for some of the states. Also, there is a bug where if you clear the screen really fast repeatedly, some of the cells "stick" and start forming patterns. Pretty cool, but totally unintended and should be fixed.