Made further progress in writing the max patch. I've decided that the sounds from the Max patch should be kept pretty simple, in contrast to the idea that we were going to use some kind of sophisticated algorithm to trigger hundreds of samples. I'm thinking more along the lines of granular synthesis, taking hints from projects like ChaOs Synth. Thinking more texturally.
I should really experiment with some sounds sometime soon, and figure out how to create/adjust them using max and the automata.

Made some graphical changes to the CAMusic program. Specifically, automata can have background colors. However, after the user creaets a mouse event, the top-left and top-right (AutomataBass and AutomataSamp2) turn red for some reason. I have not been able to locate the source of the problem yet. I expanded the Contour Automaton to cover all screen realestate (still only responds to mouse on bottom third of screen---this is not a bug). Perhaps when the "on" cells in this automaton cover most of the screen, we can end the piece.