Added dependencies that are required to run GoL to the git repository. I hope this doesn't make it too big to download.
There has been a lot of development recently, especially involving the max
patches. Most recent items have been creating a CASoundPen class that is used by
Max in order to draw / modify the automata in CAMusic. I'm not quite happy with
the implementation of this ... I feel like there are some redundant
functionalities of CASoundPen.

As always, continuing to work on the speed of CAMusic. Once it was working with
OpenGL, things started running a lot faster, but this has only been tested to
work with MacOSX. I think the way the Makefile is structured to work with OpenGL
is frankly pretty jank---this needs to be worked on. Besides OpenGL usage, there
have got to be ways to make the program faster. For example, Automata shouldn't
really have to render every single cell, just the on cells. Cells can be
maintained in a hash table, then an iterator may be used to go through every
cell in the table and render it. Much better, right?