Added many more features.
The max patch I'm working with is now "scoreViewer2". This is a much more
slimmed-down version of the original scoreViewer, which is also something of a
misnomer at this point, since it is also doing all the DSP. A more appropriate
name will be devised later.

This commit is to save the state of this project as it is. The project will now
fork in two (perhaps three) directions. They are as enumerated below:

1. A sound/video experience, primarily designed for the benefit of a
performer. Also highly entertaining to watch as an "audience member", if you
wish. Music can be largely improvised, although it is very possible a score
could be created (non-generative) that the performer could follow, as sound and
video are generated from automata.

2. A more directed performance, where suggestions (like dynamics, articulations,
pitch range) can be dictated to the performer from the computer. Other artistic
decisions are left up to the instrumentalist. Still, the view of the "score", or
whatever means is used to transmit directions to the performer should be
overlainon the automaton, rather than contained in a max patch.

(3). More or less the same direction in which the project has been going. This
means that the computer has a much more direct role in the composition of the
piece as the performer is playing it (includes rhythms, notes, perhaps
durations, etc.). This may be difficult.

I am currently most interested in implementing the first, since it will be the
easiest and (I think) the most enjoyable. Also, I have a concert coming up and
this piece seriously needs to move.